+++ profiler/manager.py
@@ -34,11 +34,24 @@ from . import helpers
 from .__version__ import __version__
 from .interface import Interface, InterfaceError
 
+__pids = []
+__pids.append(("main", os.getpid()))
+__iface = Interface()
+
 
 def signal_handler(signum, frame):
     """Handle noisy keyboardinterrupt"""
     if signum == 2:
-        print(f"profiler PID {os.getpid()} detected SIGINT or Control-C... exiting...")
+        for name, pid in __pids:
+            # We only want to print exit messages once as multiple processes close
+            if name == "main" and os.getpid() == pid:
+                if __iface.requires_monitor_interface:
+                    __iface.reset_interface()
+                    print(
+                        "Detected SIGINT or Control-C ... Removing monitor interface ... Exiting ..."
+                    )
+                else:
+                    print("Detected SIGINT or Control-C ... Exiting ...")
         sys.exit(2)
 
 
@@ -122,7 +135,6 @@ def start(args: argparse.Namespace):
             log.error("configuration validation failed... exiting...")
             sys.exit(-1)
 
-        iface_name = config.get("GENERAL").get("interface")
         listen_only = config.get("GENERAL").get("listen_only")
 
         from .fakeap import Sniffer, TxBeacons
@@ -132,20 +144,32 @@ def start(args: argparse.Namespace):
         lock = mp.Lock()
         sequence_number = mp.Value("i", 0)
 
+        iface_name = config.get("GENERAL").get("interface")
+        __iface.name = iface_name
+
         try:
             if args.no_interface_prep:
-                log.debug(
+                log.warning(
                     "user provided `--noprep` argument meaning profiler will not handle staging the interface"
                 )
-                iface = Interface(iface_name, no_interface_prep=True)
-                config["GENERAL"]["channel"] = iface.channel
+                # get channel from `iw`
+                __iface.no_interface_prep = True
+                __iface.setup()
+                if __iface.channel:
+                    config["GENERAL"]["channel"] = __iface.channel
+                log.debug("finish interface setup with no staging ...")
             else:
+                # get channel from config (CLI option or config.ini)
                 channel = int(config.get("GENERAL").get("channel"))
-                iface = Interface(iface_name, channel)
-                iface.stage_interface()
-                log.debug("finish interface prep...")
-        except InterfaceError as error:
-            log.warning("%s ... exiting ..." % error)
+                __iface.channel = channel
+                __iface.setup()
+                if __iface.requires_monitor_interface:
+                    # we require using a mon interface, update config so our subprocesses know to use it
+                    config["GENERAL"]["interface"] = __iface.mon
+                __iface.stage_interface()
+                log.debug("finish interface setup and staging ...")
+        except InterfaceError:
+            log.error("could not setup interface .... exiting ...")
             sys.exit(-1)
 
         helpers.generate_run_message(config)
@@ -163,6 +187,7 @@ def start(args: argparse.Namespace):
             )
             processes.append(txbeacons)
             txbeacons.start()
+            __pids.append(("txbeacons", txbeacons.pid))
 
         log.debug("sniffer process")
         sniffer = mp.Process(
@@ -172,6 +197,7 @@ def start(args: argparse.Namespace):
         )
         processes.append(sniffer)
         sniffer.start()
+        __pids.append(("sniffer", sniffer.pid))
 
     from .profiler import Profiler
 
@@ -179,9 +205,11 @@ def start(args: argparse.Namespace):
     profiler = mp.Process(name="profiler", target=Profiler, args=(config, queue))
     processes.append(profiler)
     profiler.start()
+    __pids.append(("profiler", profiler.pid))
 
     shutdown = False
 
+    # keep main process alive until all subprocesses are finished or closed
     while processes:
         for process in processes:
             if shutdown: